implement correction using rejection sampling
